move/rename some forge specific stuff (#4655)
