+++ src/main/java/com/gregtechceu/gtceu/api/capability/GTCapabilityHelper.java
@@ -1,6 +1,5 @@
 package com.gregtechceu.gtceu.api.capability;
 
-import com.gregtechceu.gtceu.api.capability.forge.GTCapability;
 import com.gregtechceu.gtceu.api.machine.feature.multiblock.IMaintenanceMachine;
 
 import net.minecraft.core.BlockPos;
