+++ src/main/java/com/gregtechceu/gtceu/api/recipe/ingredient/EnergyStack.java
@@ -1,7 +1,7 @@
 package com.gregtechceu.gtceu.api.recipe.ingredient;
 
 import com.gregtechceu.gtceu.api.capability.recipe.IO;
-import com.gregtechceu.gtceu.api.codec.GTCodecUtils;
+import com.gregtechceu.gtceu.utils.codec.GTCodecUtils;
 
 import net.minecraft.network.FriendlyByteBuf;
