+++ src/main/java/com/gregtechceu/gtceu/common/pipelike/optical/OpticalNetWalker.java
@@ -1,6 +1,6 @@
 package com.gregtechceu.gtceu.common.pipelike.optical;
 
-import com.gregtechceu.gtceu.api.capability.forge.GTCapability;
+import com.gregtechceu.gtceu.api.capability.GTCapability;
 import com.gregtechceu.gtceu.api.pipenet.PipeNetWalker;
 import com.gregtechceu.gtceu.common.blockentity.OpticalPipeBlockEntity;
